+++ CHANGELOG.md
@@ -29,6 +29,8 @@ All significant changes to **Pop!_OS Setup** will be documented here.
 ### Added
 - [.vimrc](samples/vimrc):
   - Added variable fg color for the vim statusline: red if you're root, white if you're not.
+- [vim.sh](post-install.d/vim.sh):
+  - Added the posibility to set vim as the default `$EDITOR`.
 ### Changed
 - [.zshrc](samples/zshrc):
   - Ubuntu style prompt (default) now has path shortening when deep in a directory structure.